feat: update simulate tx in sequencer
